More work on Commands
getAdditionalInfo works and is run as soon as a device is found (if it
is a matchstick) it runs async and when done it will fire an event to
update the device.
getAccessPoints also works and returns a list of access points, I also
have code to check if the access point is a matchstick which should work
but hasn't been tested yet
reboot works swell as well.