Lots of refactoring
It can now find all SSDP devices on the local network.  It will only
show DIAL devices but the logic is now there to allow it to show more
than just DIAL and it does find all SSDP.  It also outputs any unknown
URNs to the console when running, if you see any that aren't UUIDs you
should add them to the constants file.

I also got rid of the xml2json converter which had a problem of adding
"unknown" to the deserialized xml and replaced it with JXON.