The previous awesomeness is wired up!
To play an audio/video file:
Only file picker works right now, click browse and select a file.
Click launch to open app and set source.
Click play to begin playback.
Pause and Stop are also wired up.

Back/Forward are not yet wired. and the black bar above the buttons is a
slider for the time, which is also not yet wired up.