More Glorious Refactoring
The main.js file is getting closer to where it should be.
Added preliminary support for Fire TV Stick as I have one to test
against now.
Other minor refactoring.

Currently it isn't showing up on my MatchStick.  Not sure if this is a
problem in the extension, a problem with my MatchStick, or if the update
that came down a few days ago broke their web_sdk.