Added Firestick POC
Currently Firestick POC doesn't work through the extension but it is
able to render the webpage on the device. and play/pause the video using
the remote.  To get the POC to work you will need to run the start
server.bat. and install the amazon firestick developer app.  Then
specify the webpage inside of the app.