Mas Trabajo Para el Tabo
Next/Prev buttons in media control at top work - doesnt auto play next
in playlist yet
Randomize Works
Can choose URL or Browse for File
Clicking play inside of file picker collapses the file picker, add to
playlist does not
Both play and add to playlist add the file to the end of the playlist,
play also plays that file
When file picker is expanded the + turns to a -
Clear playlist works
Double click file in playlist to play that file
X inside of file in playlist will remove the file, and if it is playing
play the next file in the list,
Next/Prev/X will wrap around the list (click prev on first item will go
to last, next on last will go to first)
Home page is a little better, easier to see what to pick.
Clicking on button in toolbar goes straight to tab without opening the
panel first.
Stopped logging to the console when failing to parse the server response
as my receiver sends back a lot of bad responses.