Work on tab and eventing
Eventing is in rough shape.  Tried to use the fxoswebserver but it isn't
meant for extensions and I couldn't seem to get it to work.  Fell back
to httpd but I have to do some terrible things to get the response body.
Added some of the additional info back to the tab