Removed JXON
It was causing super lag on my network.  Also allows the elimination of
prefixes which are different between the xbox 360 and xbone.

Supported Types is not currently working - I tried to add in support for
the source element - I was only using sink before Source is for servers,
sink is for renderers.  Somewhere between that and removing the JXON it
doesnt properly work anymore :(

Rebuilt XPI