few bug fixes
Devices respond multiple times to a single msearch.  This leads to
requesting the device info multiple times.  Now the location of the
device info is recorded and we only go after it once during a search.

Added promises to the services
Fixed a bug in the device.html with the supported formats tabs
Moved the raw info into a collapse instead of a popover
Playlist defaults to open.