fxos web server is working for events
Can subscribe to events but can't do much with it right now
Not sure the implementation is solid -- every time it searches it tries
to subscribe again, and devices will allow the extension to subscribe
more than once meaning the same event can generate many notifications.
I have attempted to limit this but I don't think it works and I haven't
tested.