Fixed two errors
First error was expecting Content-Length to be in that specific case,
when it wasn't the server didn't think there was any content.  --> which
is why my receiver didn't seem to send a body

Second error was another spot that was expecting the data to be an
arraybuffer when it was really a string.  I fixed one of those spots
before but turns out there was another.