Updated Device/Service type/id
The device type, service type, and service id all are in a specific
format for UPnP.  I now parse the format and store the separate
segments.
Cleaned up the device/service types in the constants file
Seemed to have introduced a strange bug.  On the light I cannot switch
between services, the dimming service is always shown.  The content for
the other service is in the html and the ids seem to be correct.  It
does not matter if I start the light before firefox or after.  Other
devices seem unaffected.
Rebuilt XPI