Initial Support for subscription events has been readded
Currently navigating to a device page will subscribe to all the services
on the page.  When events are received they should be getting logged to
the console.

TODOs
Subscriptions last an hour, either set an interval to renew or extend
the timeout
Make sure services have an event url before subscribing
On $scope destruction I should unsubscribe from the services

There is an error about dupes in the capabilities tab, look into