Fixed multiple bugs see description
Fixed subscriptions, there were lots of problems, on both the front and
backend
There is an issue where sometimes it appears that the device controller
gets called twice and thus subscribes to the events twice.  This was
causing issues where angular would throw an error about dupes in the
event log.  It no longer complains about dupes but it does still
sometimes have duplicate subscriptions and I cannot tell why
Resubscribing to events still does not work
Cleaned up the event log a bit, added timestamps and service name to the
events
Removed the buffer provider as I have a better way (better in the sense
that I don't need to require an addonsdk to do it.)
Removed the chromecast and firetvstick services as they were no more
than shells that weren't doing anything anyway
Consolidated the text files
rebuilt xpi