Started the port to chrome
Changed the way I send messages to the tab to more closely align with
the way it works in chrome
Currently wont run in chrome as I have only ported it for the tab/button
APIs, you will need to run babel to convert the es6 to es5.  Also you
will need to add all of the script files inside of data directly to the
index.html page with script tags for it to work in chrome.
rebuilt xpi