Continuing work for chrome port, nonfunctional at the moment
I was in the process of adding a layer between the firefox/chrome apis
and my code to allow me to use the same code for both.  It now appears
that what I was trying to do is not possible in a chrome extension and
would have to be a chrome app instead.  Since some of the sdks I used
are for a chrome extension and some are for a chrome app I am going to
shelve this for the time being.