Fixed main problem with subscriptions and the EventService going nuts on device pages
Inside of the tab.js it was subscribing to the FrontEndBridges
sendToFrontEnd event every time it got a message from the front.  Which
means that every message from the front added an extra response for
every message going to the front.

There is still a bug where the subscriptions dont seem to be renewing
their subscription but resubscribing.